The build process for 'publish' in firebase is failing with the below er
